[Android] fix oom while uploading big files
For files larger than `Int.MAX_VALUE` the size wouldn't be returned
correctly. We have changed IPC to use `Long` for numbers instead. There
is a danger of returning a value larger than JS
`Number.MAX_SAFE_INTEGER` but it should cover most of our cases.

Kotlin's `readBytes()` takes `InputStream#available()` into account
which is not implemented for `BoundedInputStream` in a way that makes it
usable for us. After certain point it would return the rest of the file
instead of the limited size which led to `readBytes()` trying to
allocate a large slab of memory at once, which would fail with OOM.

We reimplemented `readBytes()` without taking `available()` into account
and we do not use `available()` For determining if the stream is empty.

+++ app-android/tutashared/src/main/java/de/tutao/tutashared/AndroidNativeCryptoFacade.kt
@@ -241,11 +241,12 @@ class AndroidNativeCryptoFacade(
 	@Throws(IOException::class, CryptoError::class)
 	override suspend fun aesEncryptFile(key: DataWrapper, fileUri: String, iv: DataWrapper): EncryptedFileInfo {
 		val inputStream = tempFs.fileStream(fileUri)
+		// available is int but we are encrypting only small chunks
 		val inputSize = inputStream.available()
 		val outputStream = ByteArrayOutputStream(inputSize)
 		aesEncrypt(key.data, inputStream, outputStream, iv.data, usePadding = true, useMac = true)
 		val outputUri = this.tempFs.createInMemoryFile(outputStream.toByteArray())
-		return EncryptedFileInfo(outputUri, inputSize)
+		return EncryptedFileInfo(outputUri, inputSize.toLong())
 	}
 
 	@Throws(IOException::class, CryptoError::class)
